+++ kanidmd/lib/src/filter.rs
@@ -25,8 +25,7 @@ use serde::Deserialize;
 use uuid::Uuid;
 
 use crate::be::{IdxKey, IdxKeyRef, IdxKeyToRef, IdxMeta, IdxSlope};
-use crate::identity::IdentityId;
-use crate::ldap::ldap_attr_filter_map;
+use crate::idm::ldap::ldap_attr_filter_map;
 use crate::prelude::*;
 use crate::schema::SchemaTransaction;
 use crate::value::{IndexType, PartialValue};
@@ -1561,11 +1560,17 @@ mod tests {
 
     #[test]
     fn test_lessthan_entry_filter() {
-        let e = unsafe { entry_init!(
-            ("userid", Value::new_iutf8("william")),
-            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
-            ("gidnumber", Value::Uint32(1000))
-        ).into_sealed_new() };
+        let e = unsafe {
+            entry_init!(
+                ("userid", Value::new_iutf8("william")),
+                (
+                    "uuid",
+                    Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))
+                ),
+                ("gidnumber", Value::Uint32(1000))
+            )
+            .into_sealed_new()
+        };
 
         let f_t1a = unsafe { filter_resolved!(f_lt("gidnumber", PartialValue::new_uint32(500))) };
         assert!(e.entry_match_no_index(&f_t1a) == false);
@@ -1579,17 +1584,23 @@ mod tests {
 
     #[test]
     fn test_or_entry_filter() {
-        let e = unsafe { entry_init!(
-            ("userid", Value::new_iutf8("william")),
-            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
-            ("gidnumber", Value::Uint32(1000))
-        ).into_sealed_new() };
+        let e = unsafe {
+            entry_init!(
+                ("userid", Value::new_iutf8("william")),
+                (
+                    "uuid",
+                    Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))
+                ),
+                ("gidnumber", Value::Uint32(1000))
+            )
+            .into_sealed_new()
+        };
 
         let f_t1a = unsafe {
             filter_resolved!(f_or!([
                 f_eq("userid", PartialValue::new_iutf8("william")),
                 f_eq("gidnumber", PartialValue::Uint32(1000)),
-           ]))
+            ]))
         };
         assert!(e.entry_match_no_index(&f_t1a));
 
@@ -1620,11 +1631,17 @@ mod tests {
 
     #[test]
     fn test_and_entry_filter() {
-        let e = unsafe { entry_init!(
-            ("userid", Value::new_iutf8("william")),
-            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
-            ("gidnumber", Value::Uint32(1000))
-        ).into_sealed_new() };
+        let e = unsafe {
+            entry_init!(
+                ("userid", Value::new_iutf8("william")),
+                (
+                    "uuid",
+                    Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))
+                ),
+                ("gidnumber", Value::Uint32(1000))
+            )
+            .into_sealed_new()
+        };
 
         let f_t1a = unsafe {
             filter_resolved!(f_and!([
@@ -1661,11 +1678,17 @@ mod tests {
 
     #[test]
     fn test_not_entry_filter() {
-        let e1 = unsafe { entry_init!(
-            ("userid", Value::new_iutf8("william")),
-            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
-            ("gidnumber", Value::Uint32(1000))
-        ).into_sealed_new() };
+        let e1 = unsafe {
+            entry_init!(
+                ("userid", Value::new_iutf8("william")),
+                (
+                    "uuid",
+                    Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))
+                ),
+                ("gidnumber", Value::Uint32(1000))
+            )
+            .into_sealed_new()
+        };
 
         let f_t1a =
             unsafe { filter_resolved!(f_andnot(f_eq("userid", PartialValue::new_iutf8("alice")))) };
@@ -1679,29 +1702,53 @@ mod tests {
 
     #[test]
     fn test_nested_entry_filter() {
-        let e1 = unsafe { entry_init!(
-            ("class", CLASS_PERSON.clone()),
-            ("uuid", Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))),
-            ("gidnumber", Value::Uint32(1000))
-        ).into_sealed_new() };
+        let e1 = unsafe {
+            entry_init!(
+                ("class", CLASS_PERSON.clone()),
+                (
+                    "uuid",
+                    Value::Uuid(uuid::uuid!("db237e8a-0079-4b8c-8a56-593b22aa44d1"))
+                ),
+                ("gidnumber", Value::Uint32(1000))
+            )
+            .into_sealed_new()
+        };
 
-        let e2 = unsafe { entry_init!(
-            ("class", CLASS_PERSON.clone()),
-            ("uuid", Value::Uuid(uuid::uuid!("4b6228ab-1dbe-42a4-a9f5-f6368222438e"))),
-            ("gidnumber", Value::Uint32(1001))
-        ).into_sealed_new() };
+        let e2 = unsafe {
+            entry_init!(
+                ("class", CLASS_PERSON.clone()),
+                (
+                    "uuid",
+                    Value::Uuid(uuid::uuid!("4b6228ab-1dbe-42a4-a9f5-f6368222438e"))
+                ),
+                ("gidnumber", Value::Uint32(1001))
+            )
+            .into_sealed_new()
+        };
 
-        let e3 = unsafe { entry_init!(
-            ("class", CLASS_PERSON.clone()),
-            ("uuid", Value::Uuid(uuid::uuid!("7b23c99d-c06b-4a9a-a958-3afa56383e1d"))),
-            ("gidnumber", Value::Uint32(1002))
-        ).into_sealed_new() };
+        let e3 = unsafe {
+            entry_init!(
+                ("class", CLASS_PERSON.clone()),
+                (
+                    "uuid",
+                    Value::Uuid(uuid::uuid!("7b23c99d-c06b-4a9a-a958-3afa56383e1d"))
+                ),
+                ("gidnumber", Value::Uint32(1002))
+            )
+            .into_sealed_new()
+        };
 
-        let e4 = unsafe { entry_init!(
-            ("class", CLASS_GROUP.clone()),
-            ("uuid", Value::Uuid(uuid::uuid!("21d816b5-1f6a-4696-b7c1-6ed06d22ed81"))),
-            ("gidnumber", Value::Uint32(1000))
-        ).into_sealed_new() };
+        let e4 = unsafe {
+            entry_init!(
+                ("class", CLASS_GROUP.clone()),
+                (
+                    "uuid",
+                    Value::Uuid(uuid::uuid!("21d816b5-1f6a-4696-b7c1-6ed06d22ed81"))
+                ),
+                ("gidnumber", Value::Uint32(1000))
+            )
+            .into_sealed_new()
+        };
 
         let f_t1a = unsafe {
             filter_resolved!(f_and!([
@@ -1759,11 +1806,11 @@ mod tests {
             ("class", CLASS_PERSON.clone()),
             ("class", CLASS_ACCOUNT.clone()),
             ("name", Value::new_iname("testperson1")),
-            ("uuid", Value::Uuid(uuid::uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))),
             (
-                "description",
-                Value::new_utf8s("testperson1")
+                "uuid",
+                Value::Uuid(uuid::uuid!("cc8e95b4-c24f-4d68-ba54-8bed76f63930"))
             ),
+            ("description", Value::new_utf8s("testperson1")),
             ("displayname", Value::new_utf8s("testperson1"))
         );
 
@@ -1771,11 +1818,11 @@ mod tests {
             ("class", CLASS_OBJECT.clone()),
             ("class", CLASS_PERSON.clone()),
             ("name", Value::new_iname("testperson2")),
-            ("uuid", Value::Uuid(uuid::uuid!("a67c0c71-0b35-4218-a6b0-22d23d131d27"))),
             (
-                "description",
-                Value::new_utf8s("testperson2")
+                "uuid",
+                Value::Uuid(uuid::uuid!("a67c0c71-0b35-4218-a6b0-22d23d131d27"))
             ),
+            ("description", Value::new_utf8s("testperson2")),
             ("displayname", Value::new_utf8s("testperson2"))
         );
