+++ kanidmd/lib/src/idm/mod.rs
@@ -10,6 +10,7 @@ pub mod credupdatesession;
 pub mod delayed;
 pub mod event;
 pub mod group;
+pub mod ldap;
 pub mod oauth2;
 pub mod radius;
 pub mod scim;
