+++ kanidmd/lib/src/idm/oauth2.rs
@@ -33,7 +33,6 @@ use tokio::sync::mpsc::UnboundedSender as Sender;
 use tracing::trace;
 use url::{Origin, Url};
 
-use crate::identity::IdentityId;
 use crate::idm::account::Account;
 use crate::idm::delayed::{DelayedAction, Oauth2ConsentGrant, Oauth2SessionRecord};
 use crate::idm::server::{
