+++ kanidmd/lib/src/idm/server.rs
@@ -27,7 +27,6 @@ use webauthn_rs::prelude::{Webauthn, WebauthnBuilder};
 use super::event::ReadBackupCodeEvent;
 use crate::credential::policy::CryptoPolicy;
 use crate::credential::softlock::CredSoftLock;
-use crate::identity::{AccessScope, IdentType, IdentUser, Limits};
 use crate::idm::account::Account;
 use crate::idm::authsession::AuthSession;
 use crate::idm::credupdatesession::CredentialUpdateSessionMutex;
@@ -54,7 +53,7 @@ use crate::idm::scim::{ScimSyncToken, SyncAccount};
 use crate::idm::serviceaccount::ServiceAccount;
 use crate::idm::unix::{UnixGroup, UnixUserAccount};
 use crate::idm::AuthState;
-use crate::ldap::{LdapBoundToken, LdapSession};
+use super::ldap::{LdapBoundToken, LdapSession};
 use crate::prelude::*;
 use crate::utils::{password_from_random, readable_password_from_random, uuid_from_duration, Sid};
 use crate::value::{Oauth2Session, Session};
