+++ kanidmd/lib/src/lib.rs
@@ -38,15 +38,12 @@ pub mod credential;
 pub mod entry;
 pub mod event;
 pub mod filter;
-pub mod identity;
-pub mod ldap;
 pub mod modify;
 pub mod utils;
 pub mod value;
 pub mod valueset;
 #[macro_use]
 mod plugins;
-mod access;
 pub mod idm;
 mod repl;
 pub mod schema;
@@ -68,23 +65,25 @@ pub mod prelude {
     pub use url::Url;
     pub use uuid::{uuid, Uuid};
 
+    pub use crate::be::Limits;
     pub use crate::constants::*;
     pub use crate::entry::{
         Entry, EntryCommitted, EntryInit, EntryInitNew, EntryInvalid, EntryInvalidCommitted,
         EntryInvalidNew, EntryNew, EntryReduced, EntryReducedCommitted, EntrySealed,
         EntrySealedCommitted, EntrySealedNew, EntryTuple, EntryValid,
     };
-    pub use crate::event::{CreateEvent, ExistsEvent, SearchEvent, ModifyEvent, DeleteEvent};
+    pub use crate::event::{CreateEvent, DeleteEvent, ExistsEvent, ModifyEvent, SearchEvent};
     pub use crate::filter::{
         f_and, f_andnot, f_eq, f_id, f_inc, f_lt, f_or, f_pres, f_self, f_spn_name, f_sub, Filter,
-        FilterInvalid, FC, FilterValid
+        FilterInvalid, FilterValid, FC,
     };
-    pub use crate::identity::{AccessScope, IdentType, Identity, IdentityId};
     pub use crate::idm::server::{IdmServer, IdmServerDelayed};
     pub use crate::modify::{
         m_assert, m_pres, m_purge, m_remove, Modify, ModifyInvalid, ModifyList, ModifyValid,
     };
+    pub use crate::server::access::AccessControlsTransaction;
     pub use crate::server::batch_modify::BatchModifyEvent;
+    pub use crate::server::identity::{AccessScope, IdentType, IdentUser, Identity, IdentityId};
     pub use crate::server::{
         QueryServer, QueryServerReadTransaction, QueryServerTransaction,
         QueryServerWriteTransaction,
