--- kanidmd/lib/src/server/search.rs
@@ -1,1 +1,0 @@
-
